домашка
так и не понял почему при rest теряется  this
и как из одного модуля можно вызвать функцию другого
